pyproject.toml
[project]
name = "dataset-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.4",
    "gradio>=6.10.0",
    "ipykernel>=7.2.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
